% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleanup.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean source data from NOAA}
\usage{
eq_clean_data(dataset)
}
\arguments{
\item{data}{A data frame with raw data obtained from NOAA website (see below)}
}
\value{
A data frame date & nyumeric latitude, longitude variables
}
\description{
\code{eq_clean_data} generate dates, & format latitude and longitude columns
}
\details{
The function requires raw date obtained from NOAA site
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}. It adds
a column DATE with cleaned date (Date format), transforms LATITUDE and
LONGITUDE columns as numeric objects
}
\examples{
\dontrun{
data <- readr::read_delim("signif.txt", delim = "\\t")
clean_data <- eq_clean_data(data)
}

}
